Groupshared memory is not a structured buffer, and an integer sum is not a float one
Four shaders: a groupshared array of a struct, a projective texture lookup at
ps_3_0, a multisampled load per sample index, and skinning through a typed
buffer. The projective one was clean.

A structured buffer was being found by its register number alone, and
groupshared memory is g0 where a structured buffer is t0 - both number zero. A
load from the first found the element type of the second and named members of
a float4 array, which is neither true nor compilable. Matched by what kind of
register it is as well now, the way the element type of a load already was.

And an integer instruction whose result is reinterpreted has to compute in
integers. A register holding the number 3 as a float, added to 4 as a float,
gives the float 7 - and the bits of 7.0f are not 7, so the asfloat that stores
the sum and the asint that reads it back give something else entirely. `(i + 1)
& 63` over a loop counter is the ordinary way to meet this, and vs_4_0/skin_buffer
computes the same as the original now that it is fixed. Only where the result
really is reinterpreted: a load's address is read as an integer too and has
nothing to do with what the texel it fetches is kept as.

Two known differences and two round trip costs recorded. Both differences are
the denormal fold again, which nine shaders now wait on.

+++ HlslDecompiler.Tests/ShaderSources/ps_4_0_instruction/int_float_mix.fx
@@ -15,7 +15,7 @@ float4 main(PS_IN i) : SV_Target
 	r1 = -i.texcoord1.xyw;
 	r0.y = (asint(r0.y) != 0) ? r1.y : i.texcoord1.y;
 	r0.zw = asfloat(max(r1.xz, i.texcoord1.xw));
-	r0.x = asfloat(r0.y + asint(r0.x));
+	r0.x = asfloat((int)r0.y + asint(r0.x));
 	r0.x = (float)asint(r0.x);
 	r0.y = frac(i.texcoord.x);
 	r1.x = (float)(int)i.texcoord1.z;
@@ -37,7 +37,7 @@ float4 main(PS_IN i) : SV_Target
 	r0.x = asfloat((int)r0.x);
 	r0.x = asfloat(asint(r0.x) + i.texcoord1.y);
 	o.x = (float)asint(r0.x);
-	r0.x = asfloat(-r0.y);
+	r0.x = asfloat(-(int)r0.y);
 	r0.z = asfloat(i.texcoord1.x ^ 2);
 	r0.z = asfloat(asint(r0.z) & -2147483648);
 	r0.x = (asint(r0.z) != 0) ? r0.x : r0.y;
